+++ Runtime/Configuration/AssemblyInfo.cs
@@ -9,6 +9,8 @@ using UnityEngine.Scripting;
 // Test assemblies
 #if UNITY_INCLUDE_TESTS
 [assembly: InternalsVisibleTo("Unity.Services.Core.Tests")]
-[assembly: InternalsVisibleTo("Unity.Services.Core.EditorTests")]
+[assembly: InternalsVisibleTo("Unity.Services.Core.TestUtils")]
+[assembly: InternalsVisibleTo("Unity.Services.Core.TestUtils.Tests")]
+[assembly: InternalsVisibleTo("Unity.Services.Core.TestUtils.EditorTests")]
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]
 #endif