+++ Runtime/Core.Internal/AssemblyInfo.cs
@@ -18,6 +18,7 @@ using System.Runtime.CompilerServices;
 // Test assemblies
 #if UNITY_INCLUDE_TESTS
 [assembly: InternalsVisibleTo("Unity.Services.Core.Tests")]
+[assembly: InternalsVisibleTo("Unity.Services.Core.TestUtils")]
 
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]
 [assembly: InternalsVisibleTo("Unity.Services.Core.EditorTests")]