+++ Runtime/Core.Internal/UnityServicesInitializer.cs
@@ -7,6 +7,7 @@ namespace Unity.Services.Core.Internal
         [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.AfterAssembliesLoaded)]
         static void CreateStaticInstance()
         {
+            CoreRegistry.Instance = new CoreRegistry();
             UnityServices.Instance = new UnityServicesInternal(CoreRegistry.Instance);
             UnityServices.InstantiationCompletion?.TrySetResult(null);
         }