+++ Runtime/Registration/AssemblyInfo.cs
@@ -6,4 +6,5 @@ using UnityEngine.Scripting;
 #if UNITY_INCLUDE_TESTS
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]
 [assembly: InternalsVisibleTo("Unity.Services.Core.Tests")]
+[assembly: InternalsVisibleTo("Unity.Services.Core.TestUtils.Tests")]
 #endif