+++ Runtime/Telemetry/TelemetryEvents/MetricType.cs
@@ -1,0 +1,9 @@
+namespace Unity.Services.Core.Telemetry.Internal
+{
+    enum MetricType
+    {
+        Observer = 0,
+        Counter = 1,
+        Timed = 2,
+    }
+}