+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0-pre.9] - 2021-09-23
+### Added
+- New common error codes: ApiMissing, RequestRejected, NotFound, InvalidRequest.
+- Link project pop-up dialog
+
+### Fixed
+- Core registry throwing exceptions when domain reloads are disabled
+
 ## [1.1.0-pre.8] - 2021-08-04
 ### Added
 - Added base exception type for other Operate SDKs to derive from. Consistent error handling experience.